refactor: improve import source validation and model upload
- Extract validateSourceUrl to importSourceUtil for better separation of concerns
- Add Civitai and Hugging Face to i18n untranslated names
- Refactor UploadModelUrlInputCivitai to use defineModel
- Add early validation check in uploadModel for better error handling
- Remove optional chaining where source is guaranteed to exist
