+++ src/platform/assets/types/importSource.ts
@@ -22,17 +22,3 @@ export interface ImportSource {
    */
   readonly hostnames: readonly string[]
 }
-
-/**
- * Check if a URL belongs to a specific import source
- */
-export function validateSourceUrl(url: string, source: ImportSource): boolean {
-  try {
-    const hostname = new URL(url).hostname.toLowerCase()
-    return source.hostnames.some(
-      (h) => hostname === h || hostname.endsWith(`.${h}`)
-    )
-  } catch {
-    return false
-  }
-}
